+++ AdminWk4/src/main/java/com/lms/LMSAdmin/dao/BorrowerDao.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lms.LMSAdmin.pojo.Borrower;
 
 @Repository
-public interface BorrowerDao extends JpaRepository<Borrower, Long> {
+public interface BorrowerDao extends JpaRepository<Borrower, Integer> {
 
 }
