+++ AdminWk4/src/main/java/com/lms/LMSAdmin/dao/LibraryBranchDao.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lms.LMSAdmin.pojo.LibraryBranch;
 
 @Repository
-public interface LibraryBranchDao extends JpaRepository<LibraryBranch, Long>{
+public interface LibraryBranchDao extends JpaRepository<LibraryBranch, Integer>{
 
 }
